backend/requirements.txt
# Production Dependencies for Hackathon II - Phase 2
# Backend FastAPI Application

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlmodel==0.0.22
psycopg2-binary==2.9.10

# Authentication & Security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4

# Configuration
python-dotenv==1.0.1
pydantic[email]==2.10.5
pydantic-settings==2.6.1

# Deployment
gunicorn==23.0.0
